Update README.md
SEO Weepers is a repository dedicated to the trials and tribulations of SEO professionals. From algorithm-induced ranking drops to the challenges of technical SEO, this repo collects stories, insights, and best practices to help navigate the ever-changing SEO landscape. Share your experiences, learn from others, and find humor in the world of search engine optimization!